document-parsing-service/requirements.txt
fastapi
uvicorn
pydantic
requests
boto3
PyMuPDF
python-docx